# docs/admin-api/admin_swagger.yaml
basePath: /api/v1
definitions:
  api.PageData:
    properties:
      list: {}
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
      requestId:
        type: string
    type: object
  dto.Base64ImageUploadReq:
    properties:
      image:
        type: string
    type: object
  dto.GetAuthorizeUrlResp:
    properties:
      url:
        type: string
    type: object
...
  dto.KfAccountCreateReq:
    properties:
      account:
        type: string
      appId:
        type: string
      nickname:
        type: string
    type: object
  dto.KfAccountDeleteReq:
    properties:
      account:
        type: string
      appId:
        type: string
    type: object
  dto.KfAccountGetAllReq:
    properties:
      appId:
        type: string
    type: object
  dto.KfAccountSetDefaultReq:
    properties:
      account:
        type: string
      appId:
        type: string
    type: object
  dto.MenuClearReq:
    properties:
      appId:
        type: string
    type: object
...
  dto.UploadFileResp:
    properties:
      url:
        type: string
    type: object
  dto.WxAuthorizerPageQueryReq:
    properties:
      keyword:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
    type: object
...
  models.WxAuthorizer:
    properties:
      authorizerAppid:
        type: string
      componentAppid:
        type: string
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      defaultKfAccount:
        type: string
      headImg:
        type: string
      id:
        example: "0"
        type: string
      nickName:
        type: string
      principalName:
        type: string
      qrcodeUrl:
        type: string
      serviceType:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
      userName:
        type: string
      verifyType:
        type: string
    type: object
host: 127.0.0.1:10000
info:
  contact:
    email: xxx@qq.com
    name: xxx
    url: http://xxx.com
  description: xxx管理后台API文档
  title: Admin API
  version: 1.0.0
paths:
...
  /file/upload_base64_image:
    post:
      consumes:
      - application/json
      description: 上传Base64格式图片
      parameters:
      - description: base64格式图片
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.Base64ImageUploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UploadFileResp'
              type: object
      summary: 上传Base64格式图片
      tags:
      - 文件上传
...
  /file/upload_image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UploadFileResp'
              type: object
      summary: 上传图片
      tags:
      - 文件上传
...
  /wx/authorizer/create_kf_account:
    post:
      consumes:
      - application/json
      description: 为公众号创建客服账号
      parameters:
      - description: 客服账号和昵称
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.KfAccountCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 创建客服账号
      tags:
      - 授权公众号管理
  /wx/authorizer/delete_kf_account:
    post:
      consumes:
      - application/json
      description: 删除客服账号
      parameters:
      - description: 公众号appId
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.KfAccountDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 删除客服账号
      tags:
      - 授权公众号管理
  /wx/authorizer/delete_menu:
    post:
      consumes:
      - application/json
      description: 清空菜单
      parameters:
      - description: 公众号appId
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.MenuClearReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 清空菜单
      tags:
      - 授权公众号管理
  /wx/authorizer/get_all_kf_accounts:
    post:
      consumes:
      - application/json
      description: 获取公众号的所有客服账号
      parameters:
      - description: 公众号appId
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.KfAccountGetAllReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 获取公众号的所有客服账号
      tags:
      - 授权公众号管理
  /wx/authorizer/get_authorize_url:
    post:
      consumes:
      - application/json
      description: 获取授权链接，微信扫码访问该连接，会跳转到选择公众号授权给第三方平台页面
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetAuthorizeUrlResp'
              type: object
      security:
      - RequireLogin: []
      summary: 获取公众号授权连接
      tags:
      - 授权公众号管理
  /wx/authorizer/get_menu:
    post:
      consumes:
      - application/json
      description: 获取菜单
      parameters:
      - description: 公众号appId
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.MenuClearReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 获取菜单
      tags:
      - 授权公众号管理
  /wx/authorizer/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询授权公众号
      parameters:
      - description: 授权公众号筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.WxAuthorizerPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.WxAuthorizer'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询授权公众号
      tags:
      - 授权公众号管理
  /wx/authorizer/set_default_kf_account:
    post:
      consumes:
      - application/json
      description: 设置默认的客服账号（使用该账号给用户发送消息）
      parameters:
      - description: 公众号appId
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.KfAccountSetDefaultReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 设置默认的客服账号（使用该账号给用户发送消息）
      tags:
      - 授权公众号管理
...
securityDefinitions:
  RequireLogin:
    in: header
    name: token
    type: apiKey
swagger: "2.0"